backend/requirements.txt
fastapi
uvicorn[standard]
zep-python>=2.0.0
openai
python-dotenv
sqlalchemy
pydantic
httpx
alembic
psycopg2-binary
